Move VenueHistory scss into separate file and work on styling
