+++ src/components/VenueHistory/VenueHistory.scss
@@ -1,0 +1,48 @@
+@import '../../styles/variables.scss';
+
+.venue-history {
+  h3 {
+    margin-bottom: 0;
+    font-weight: 300;
+    font-size: 1.3rem;
+  }
+
+  .city {
+    color: $pink-four;
+  }
+
+  .event-details {
+    font-size: 1.5em;
+  }
+
+  .event-container {
+    max-height: 140px;
+    overflow: scroll;
+    background-color: #32333e;
+  }
+
+  .table-headings {
+    display: grid;
+    grid-template-columns: repeat(3, 10fr);
+    border-bottom: 1px solid #fff;
+    align-items: center;
+    padding-left: 2rem;
+    margin: 0;
+
+    h4 {
+      margin-bottom: 5px;
+    }
+  }
+
+  .table-row {
+    display: grid;
+    grid-template-columns: repeat(3, 10fr);
+    align-items: center;
+    justify-content: center;
+
+    margin-left: 2rem;
+
+    font-weight: 300;
+    font-size: 0.6rem;
+  }
+}
